+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/Display.kt
@@ -9,5 +9,4 @@ interface Display {
     fun getDisplayInstalledVersion(context: Context): String
     val displayDownloadSource: Int
     val displayIcon: Int
-    val displayIconBackground: Int
 }