+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/FirefoxNightly.kt
@@ -2,7 +2,6 @@ package de.marmaro.krt.ffupdater.app.impl
 
 import android.content.Context
 import android.content.pm.PackageManager
-import android.graphics.Color
 import android.os.Build
 import androidx.preference.PreferenceManager
 import de.marmaro.krt.ffupdater.R
@@ -27,7 +26,6 @@ class FirefoxNightly(private val apiConsumer: ApiConsumer) : BaseAppDetail() {
     override val displayWarning = R.string.firefox_nightly__warning
     override val displayDownloadSource = R.string.mozilla_ci
     override val displayIcon = R.mipmap.ic_logo_firefox_nightly
-    override val displayIconBackground = Color.parseColor("#FFFFFF")
     override val minApiLevel = Build.VERSION_CODES.LOLLIPOP
     override val supportedAbis = listOf(ABI.ARM64_V8A, ABI.ARMEABI_V7A, ABI.X86_64, ABI.X86)