+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/FirefoxRelease.kt
@@ -1,6 +1,5 @@
 package de.marmaro.krt.ffupdater.app.impl
 
-import android.graphics.Color
 import android.os.Build
 import de.marmaro.krt.ffupdater.R
 import de.marmaro.krt.ffupdater.app.AvailableVersionResult
@@ -20,7 +19,6 @@ class FirefoxRelease(private val apiConsumer: ApiConsumer) : BaseAppDetail() {
     override val displayWarning: Int? = null
     override val displayDownloadSource = R.string.mozilla_ci
     override val displayIcon = R.mipmap.ic_logo_firefox_release
-    override val displayIconBackground = Color.parseColor("#FFFFFF")
     override val minApiLevel = Build.VERSION_CODES.LOLLIPOP
     override val supportedAbis = listOf(ABI.ARM64_V8A, ABI.ARMEABI_V7A, ABI.X86_64, ABI.X86)