+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/Iceraven.kt
@@ -1,7 +1,6 @@
 package de.marmaro.krt.ffupdater.app.impl
 
 import android.content.Context
-import android.graphics.Color
 import android.os.Build
 import de.marmaro.krt.ffupdater.R
 import de.marmaro.krt.ffupdater.app.AvailableVersionResult
@@ -23,7 +22,6 @@ class Iceraven(private val apiConsumer: ApiConsumer) : BaseAppDetail() {
     override val displayWarning: Int? = null
     override val displayDownloadSource = R.string.github
     override val displayIcon = R.mipmap.ic_logo_iceraven
-    override val displayIconBackground = Color.parseColor("#FFFFFF")
     override val minApiLevel = Build.VERSION_CODES.LOLLIPOP
     override val supportedAbis = listOf(ABI.ARM64_V8A, ABI.ARMEABI_V7A, ABI.X86_64, ABI.X86)